Refactor component imports and add new About section components
