+++ components/landing/DiscountModal.tsx
