+++ components/landing/HeroBackground.tsx
